+++ src/interrupt_signal.rs
@@ -7,3 +7,18 @@ impl From<((), InterruptSignal)> for InterruptSignal {
         InterruptSignal
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::InterruptSignal;
+
+    #[test]
+    fn debug() {
+        assert_eq!("InterruptSignal", format!("{:?}", InterruptSignal))
+    }
+
+    #[test]
+    fn clone() {
+        assert_eq!(InterruptSignal, Clone::clone(&InterruptSignal))
+    }
+}